python_architecture_linter_cli/pyproject.toml
[tool.poetry]
name = "python_architecture_linter_cli"
version = "0.1.3"
description = "CLI to Lint architecture"
authors = ["Brian Graham <Brian@statagroup.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.0.1"
"ruamel.yaml" = "^0.17.10"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
flake8 = "^3.9.2"
mypy = "^0.902"
black = "^21.6b0"
isort = "^5.8.0"
coverage = "^5.5"
pytest-cov = "^2.12.1"
